backend/requirements.txt
Flask
Flask-Session
Flask-Mail
python-dotenv
reportlab
bcrypt
requests
Flask-WTF
Flask-Limiter
